rename configs --> recipes
