+++ nemo/collections/llm/recipes/llama2_7b.py
@@ -2,8 +2,8 @@ import pytorch_lightning as pl
 
 from nemo import lightning as nl
 from nemo.collections.llm.api import finetune, pretrain
-from nemo.collections.llm.configs.log.default import default_log
-from nemo.collections.llm.configs.optim.adam import adam_with_cosine_annealing
+from nemo.collections.llm.recipes.log.default import default_log
+from nemo.collections.llm.recipes.optim.adam import adam_with_cosine_annealing
 from nemo.collections.llm.gpt.data.api import squad
 from nemo.collections.llm.gpt.model.llama import Llama2Config7B, LlamaModel
 from nemo.collections.llm.peft.api import gpt_lora
