+++ nemo/collections/llm/recipes/llama3_8b_64k.py
@@ -2,8 +2,8 @@ import pytorch_lightning as pl
 
 from nemo import lightning as nl
 from nemo.collections.llm.api import pretrain
-from nemo.collections.llm.configs.log.default import default_log
-from nemo.collections.llm.configs.optim.adam import adam_with_cosine_annealing
+from nemo.collections.llm.recipes.log.default import default_log
+from nemo.collections.llm.recipes.optim.adam import adam_with_cosine_annealing
 from nemo.collections.llm.gpt.data.api import squad
 from nemo.collections.llm.gpt.model.llama import Llama3Config8B, LlamaModel
 from nemo.collections.llm.utils import Partial, factory
